pyproject.toml
[project]
name = "infrasentinel"
version = "0.1.0"
description = "InfraSentinel monorepo scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "pydantic-ai>=0.0.14",
  "temporalio>=1.7.0",
  "sqlalchemy>=2.0.34",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "alembic>=1.13.2",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "boto3>=1.35.0",
  "mcp>=1.2.0",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.9",
  "mypy>=1.11.2",
  "pre-commit>=3.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
plugins = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
